Refactor code to extract download_file function
