+++ FileDownloader.py
@@ -1,0 +1,4 @@
+from Google import create_service
+from OrderHandler import resource_path
+
+DRIVES_PATH = resource_path("drive_folders_destination.json")
