+++ Google.py
@@ -1,45 +1,43 @@
 import pickle
 import os
-from google_auth_oauthlib.flow import Flow, InstalledAppFlow
+
+from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
-from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
 from google.auth.transport.requests import Request
+from datetime import datetime
 
 
-def Create_Service(client_secret_file, api_name, api_version, *scopes):
-    print(client_secret_file, api_name, api_version, scopes, sep='-')
-    CLIENT_SECRET_FILE = client_secret_file
-    API_SERVICE_NAME = api_name
-    API_VERSION = api_version
-    SCOPES = [scope for scope in scopes[0]]
-    print(SCOPES)
+def create_service(client_secret_file, api_name, api_version, *scopes):
+    print(client_secret_file, api_name, api_version, scopes, sep="-")
+    scopes = [scope for scope in scopes[0]]
 
     cred = None
 
-    pickle_file = f'token_{API_SERVICE_NAME}_{API_VERSION}.pickle'
+    pickle_file = f"token_{api_name}_{api_version}.pickle"
     # print(pickle_file)
 
     if os.path.exists(pickle_file):
-        with open(pickle_file, 'rb') as token:
+        with open(pickle_file, "rb") as token:
             cred = pickle.load(token)
 
     if not cred or not cred.valid:
         if cred and cred.expired and cred.refresh_token:
             cred.refresh(Request())
         else:
-            flow = InstalledAppFlow.from_client_secrets_file(CLIENT_SECRET_FILE, SCOPES)
+            flow = InstalledAppFlow.from_client_secrets_file(client_secret_file, scopes)
             cred = flow.run_local_server()
 
-        with open(pickle_file, 'wb') as token:
+        with open(pickle_file, "wb") as token:
             pickle.dump(cred, token)
 
     try:
-        service = build(API_SERVICE_NAME, API_VERSION, credentials=cred, static_discovery=False)
-        print(API_SERVICE_NAME, 'service created successfully')
+        service = build(api_name, api_version, credentials=cred, static_discovery=False)
+        print(api_name, "service created successfully")
         return service
     except Exception as e:
         raise Exception(e)
 
+
 def convert_to_RFC_datetime(year=1900, month=1, day=1, hour=0, minute=0):
-    dt = datetime.datetime(year, month, day, hour, minute, 0).isoformat() + 'Z'
+    dt = datetime.datetime(year, month, day, hour, minute, 0).isoformat() + "Z"
     return dt
