+++ constants.py
@@ -1,0 +1,17 @@
+import json
+
+from utils import resource_path
+
+
+SMALL_SIZES = ["3-4", "5-6", "7-8"]
+PRODUCTS = ["LEZAK", "KOSZ", "POD", "KB_ZW", "KB_MAG", "KB_FUN"]
+
+DRIVES_PATH = resource_path("drive_folders_destination.json")
+with open(DRIVES_PATH) as f:
+    drive_dest = json.load(f)
+
+    WHITE_SHIRT_FOLDER_ID = drive_dest.get("WHITE_SHIRT_FOLDER_ID")
+    WHITE_CUP_FOLDER_ID = drive_dest.get("WHITE_CUP_FOLDER_ID")
+    BLACK_SHIRT_FOLDER_ID = drive_dest.get("BLACK_SHIRT_FOLDER_ID")
+    BLACK_CUP_FOLDER_ID = drive_dest.get("BLACK_CUP_FOLDER_ID")
+    BLACK_HALFTONE_SHIRT_FOLDER_ID = drive_dest.get("BLACK_HALFTONE_SHIRT_FOLDER_ID")
